+++ src/modules/dforest/dforestsnv.h
@@ -22,7 +22,7 @@ public:
     
     bool run(string encode_file, string align_file, string cmpreads_file, string a_out_file, int min_reads, int max_depth);
     
-    void build_tree(const vector<int> &cand_loci, long long &counter, vector<long long> &temp_vec_var, vector<long long> &temp_vec_read, int min_reads, int max_depth);
+    void build_tree(const vector<int> &cand_loci, int64_t &counter, vector<int64_t> &temp_vec_var, vector<int64_t> &temp_vec_read, int min_reads, int max_depth);
 };
